Move CookieComparer to common
We need to use this in HttpListenerRequest.Cookies as we need to add a
custom implementation of CookieCollection.InternalAdd(cookie, true)
