+++ src/System.Net.Primitives/src/System.Net.Primitives.csproj
@@ -62,6 +62,9 @@
     <Compile Include="$(CommonPath)\System\Net\ByteOrder.cs">
       <Link>Common\System\Net\ByteOrder.cs</Link>
     </Compile>
+    <Compile Include="$(CommonPath)\System\Net\CookieComparer.cs">
+      <Link>ProductionCode\System\Net\CookieComparer.cs</Link>
+    </Compile>
     <Compile Include="$(CommonPath)\System\Net\CookieFields.cs">
       <Link>Common\System\Net\CookieFields.cs</Link>
     </Compile>
